refactor: 네트워크 공통 로직 network/common.ts로 이동
