+++ actions/seminarServer.ts
@@ -1,7 +1,7 @@
 import { PostSearchQueryParams } from '@/types/post';
 import { SeminarList, Seminar } from '@/types/seminar';
 
-import { deleteRequest, getRequest } from '../apis/common/server';
+import { deleteRequest, getRequest } from '../apis/network/server';
 
 const seminarPath = '/seminar';
