+++ apis/main.ts
@@ -1,5 +1,5 @@
 import { MainContents } from '@/types/main';
 
-import { getRequest } from './common/client';
+import { getRequest } from './network/client';
 
 export const getMainContents = () => getRequest('') as Promise<MainContents>;
