+++ apis/network/client.ts
@@ -1,11 +1,6 @@
 import { objToQueryString } from '@/utils/convertParams';
 
-// export const BASE_URL =
-//   process.env.NODE_ENV === 'production'
-//     ? 'https://cse-dev-waffle.bacchus.io/api/v1'
-//     : 'http://localhost:8080/api/v1';
-
-export const BASE_URL = 'https://cse-dev-waffle.bacchus.io/api/v1';
+import { BASE_URL, checkError } from './common';
 
 export const getRequest = async <T = unknown>(
   url: string,
@@ -47,17 +42,3 @@ export const deleteRequest = async (url: string, init?: RequestInit) => {
   const response = await fetch(fetchUrl, { ...init, method: 'DELETE', credentials: 'include' });
   checkError(response);
 };
-
-export const checkError = (response: Response) => {
-  if (response.ok) return;
-
-  throw new NetworkError(response.status);
-};
-
-export class NetworkError extends Error {
-  statusCode: number;
-  constructor(statusCode: number) {
-    super(`${statusCode} 에러`);
-    this.statusCode = statusCode;
-  }
-}
