+++ constants/network.ts
@@ -1,1 +1,5 @@
+import { BASE_URL } from '@/apis/network/common';
+
 export const COOKIE_SESSION_ID = 'JSESSIONID';
+export const LOGIN_URL = BASE_URL + '/login';
+export const LOGOUT_URL = BASE_URL + '/logout';
